backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Elevator Advocacy Platform Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django~=6.0",
    "djangorestframework",
    "requests",
    "python-dotenv",
    "django-cors-headers>=4.9.0",
    "google-genai>=1.72.0",
    "instructor>=1.15.1",
    "serpapi>=1.0.2",
    "types-requests>=2.31.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]
